+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [3.5.2] 2020-04-20
+
 -   Summarize time-field correctly
 -   Report errors and failures separately